Tweak Organization View
Rather than display all organizations, just show the one that's in scope
as chosen by the app bar as it makes a lot more sense 99% of the time.
