+++ charts/ui/Chart.yaml
@@ -4,8 +4,8 @@ description: A Helm chart for deploying Unikorn UI
 
 type: application
 
-version: v0.2.12
-appVersion: v0.2.12
+version: v0.2.13
+appVersion: v0.2.13
 
 icon: https://assets.unikorn-cloud.org/assets/images/logos/dark-on-light/icon.png
